gif_generator/requirements.txt

openai-whisper
moviepy
